requirements.txt
requests>=2.31.0
ddddocr>=1.4.11
pycryptodome>=3.19.0
tqdm>=4.66.1